Collapse down to one runner
